+++ pytsite/assetman/_console_command.py
@@ -26,7 +26,7 @@ class Build(_console.command.Abstract):
     def get_options_help(self) -> str:
         """Get help for the command.
         """
-        return '[--package=NAME] [--no-maint]'
+        return '[--package=NAME] [--no-maint] [--no-cache]'
 
     def get_options(self) -> tuple:
         """Get command options.
@@ -34,13 +34,14 @@ class Build(_console.command.Abstract):
         return (
             ('package', _validation.rule.Pass()),
             ('no-maint', _validation.rule.Pass()),
+            ('no-cache', _validation.rule.Pass()),
         )
 
     def execute(self, args: tuple = (), **kwargs):
         """Execute The Command.
         """
         try:
-            _api.build(kwargs.get('package'), not kwargs.get('no-maint'))
+            _api.build(kwargs.get('package'), not kwargs.get('no-maint'), not kwargs.get('no-cache'))
 
         except (_error.PackageNotRegistered, _error.PackageAlreadyRegistered) as e:
             raise _console.error.Error(e)